LPD-85361 Refactor: separate bulk actions in new file
